+++ src/helpers.js
@@ -1,5 +1,9 @@
 const PLUGIN_ENVIRONMENT_VAR = "FAIL_FAST";
+const SHOULD_SKIP_TASK = "failFastShouldSkip";
+const RESET_SKIP_TASK = "failFastResetSkip";
 
 module.exports = {
   PLUGIN_ENVIRONMENT_VAR,
+  SHOULD_SKIP_TASK,
+  RESET_SKIP_TASK,
 };