+++ test/plugin.spec.js
@@ -5,14 +5,14 @@ describe("plugin", () => {
   let onEventSpy;
   let sandbox;
   let eventName;
-  let shouldSkipDueToFailFast;
-  let resetShouldSkipDueToFailFast;
+  let failFastShouldSkip;
+  let failFastResetSkip;
 
   const getPluginMethods = () => {
     plugin(onEventSpy);
     eventName = onEventSpy.getCall(0).args[0];
-    shouldSkipDueToFailFast = onEventSpy.getCall(0).args[1].shouldSkipDueToFailFast;
-    resetShouldSkipDueToFailFast = onEventSpy.getCall(0).args[1].resetShouldSkipDueToFailFast;
+    failFastShouldSkip = onEventSpy.getCall(0).args[1].failFastShouldSkip;
+    failFastResetSkip = onEventSpy.getCall(0).args[1].failFastResetSkip;
   };
 
   beforeEach(() => {
@@ -34,55 +34,55 @@ describe("plugin", () => {
   describe("when it is disabled", () => {
     it("should return shouldSkip flag as false by default", () => {
       getPluginMethods();
-      expect(shouldSkipDueToFailFast()).toEqual(false);
+      expect(failFastShouldSkip()).toEqual(false);
     });
 
     it("should return shouldSkip flag as true after setting it as true", () => {
       getPluginMethods();
-      shouldSkipDueToFailFast(true);
-      expect(shouldSkipDueToFailFast()).toEqual(true);
+      failFastShouldSkip(true);
+      expect(failFastShouldSkip()).toEqual(true);
     });
   });
 
   describe("when it is enabled using boolean", () => {
     it("should return shouldSkip flag as false by default when plugin ", () => {
       getPluginMethods();
-      expect(shouldSkipDueToFailFast()).toEqual(false);
+      expect(failFastShouldSkip()).toEqual(false);
     });
 
     it("should return shouldSkip flag as true after setting it as true", () => {
       getPluginMethods();
-      shouldSkipDueToFailFast(true);
-      expect(shouldSkipDueToFailFast()).toEqual(true);
+      failFastShouldSkip(true);
+      expect(failFastShouldSkip()).toEqual(true);
     });
 
     it("should reset shouldSkip flag after calling to reset method", () => {
       getPluginMethods();
-      shouldSkipDueToFailFast(true);
-      expect(shouldSkipDueToFailFast()).toEqual(true);
-      resetShouldSkipDueToFailFast();
-      expect(shouldSkipDueToFailFast()).toEqual(false);
+      failFastShouldSkip(true);
+      expect(failFastShouldSkip()).toEqual(true);
+      failFastResetSkip();
+      expect(failFastShouldSkip()).toEqual(false);
     });
   });
 
   describe("when it is enabled using string", () => {
     it("should return shouldSkip flag as false by default when plugin ", () => {
       getPluginMethods();
-      expect(shouldSkipDueToFailFast()).toEqual(false);
+      expect(failFastShouldSkip()).toEqual(false);
     });
 
     it("should return shouldSkip flag as true after setting it as true", () => {
       getPluginMethods();
-      shouldSkipDueToFailFast(true);
-      expect(shouldSkipDueToFailFast()).toEqual(true);
+      failFastShouldSkip(true);
+      expect(failFastShouldSkip()).toEqual(true);
     });
 
     it("should reset shouldSkip flag after calling to reset method", () => {
       getPluginMethods();
-      shouldSkipDueToFailFast(true);
-      expect(shouldSkipDueToFailFast()).toEqual(true);
-      resetShouldSkipDueToFailFast();
-      expect(shouldSkipDueToFailFast()).toEqual(false);
+      failFastShouldSkip(true);
+      expect(failFastShouldSkip()).toEqual(true);
+      failFastResetSkip();
+      expect(failFastShouldSkip()).toEqual(false);
     });
   });
 });